+++ packages/js/src/Modules/Verto/webrtc/Peer.ts
@@ -254,12 +254,10 @@ export default class Peer {
       streams: [first],
     } = event;
     const { remoteElement, screenShare } = this.options;
-    let { remoteStream } = this.options;
-
-    remoteStream = first;
+    this.options.remoteStream = first;
 
     if (screenShare === false) {
-      attachMediaStream(remoteElement, remoteStream);
+      attachMediaStream(remoteElement, this.options.remoteStream);
     }
   }