+++ packages/js/CHANGELOG.md
@@ -4,6 +4,7 @@
 - chore: restart debug stats reporter on session reconnect
 - chore: expose PeerConnectionSignalingStateClosed event
 - chore: update documentation for event and error handling
+- fix: remote stream assignment to properly persist (@maiconpavi)
 
 ## [2.25.14](https://github.com/team-telnyx/webrtc/compare/webrtc/v2.25.13...webrtc/v2.25.14) (2025-12-16)